refactor(api.py): remove background refresh task from Otf class to simplify class responsibilities fix(auth.py): move background refresh task to OtfUser class to ensure token refresh logic is encapsulated within the user class feat(auth.py): add validation for refresh_callback to ensure it is a callable function with one argument
